requirements.txt
asgiref==3.3.4
Django==3.2.2
django-crispy-forms==1.11.2
django-debug-toolbar==3.2.1
django-taggit==1.4.0
Pillow==8.2.0
psycopg2==2.8.6
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
